src/requirements.txt
matplotlib
pandas
pillow
wordcloud